# apps/backend/src/assets/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
    "/stock/{symbol}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SearchResults",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "AAPL"
                    },
                    "sma": {
                      "type": "number",
                      "example": 173.23
                    },
                    "ema": {
                      "type": "number",
                      "example": 172.778916037368
                    },
                    "current": {
                      "type": "number",
                      "example": 173.23
                    },
                    "updated": {
                      "type": "string",
                      "example": "2024-03-13T10:31:00.663Z"
                    },
                    "currency": {
                      "type": "string",
                      "example": "USD"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "AAPL"
                    },
                    "sma": {
                      "type": "number",
                      "example": 173.23
                    },
                    "ema": {
                      "type": "number",
                      "example": 172.778916037368
                    },
                    "current": {
                      "type": "number",
                      "example": 173.23
                    },
                    "updated": {
                      "type": "string",
                      "example": "2024-03-13T10:31:00.663Z"
                    },
                    "currency": {
                      "type": "string",
                      "example": "USD"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search/{symbol}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SearchResults",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "description": {
                        "type": "string",
                        "example": "APPLE INC"
                      },
                      "displaySymbol": {
                        "type": "string",
                        "example": "AAPL"
                      },
                      "symbol": {
                        "type": "string",
                        "example": "AAPL"
                      },
                      "type": {
                        "type": "string",
                        "example": "Common Stock"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "description": {
                        "type": "string",
                        "example": "APPLE INC"
                      },
                      "displaySymbol": {
                        "type": "string",
                        "example": "AAPL"
                      },
                      "symbol": {
                        "type": "string",
                        "example": "AAPL"
                      },
                      "type": {
                        "type": "string",
                        "example": "Common Stock"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/chart/{symbol}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ChartResult",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 219
                      },
                      "stockId": {
                        "type": "number",
                        "example": 2
                      },
                      "price": {
                        "type": "number",
                        "example": 173.23
                      },
                      "ema": {
                        "type": "number",
                        "example": 172.8027945757707
                      },
                      "sma": {
                        "type": "number",
                        "example": 173.23
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-03-13T10:26:00.442Z"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 219
                      },
                      "stockId": {
                        "type": "number",
                        "example": 2
                      },
                      "price": {
                        "type": "number",
                        "example": 173.23
                      },
                      "ema": {
                        "type": "number",
                        "example": 172.8027945757707
                      },
                      "sma": {
                        "type": "number",
                        "example": 173.23
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-03-13T10:26:00.442Z"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SearchResult": {
        "type": "object",
        "properties": {},
        "xml": {
          "name": "SearchResult"
        }
      }
    }
  },
  "swaggerDefinition": {
    "restapi": "1.0.0",
    "info": {
      "title": "Stock price Checker API",
      "version": "1.0.0",
      "description": "It's a price checker API, really"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ]
  },
  "apis": [
    "**/*.js"
  ]
}